% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{summary.split_df}
\alias{summary.split_df}
\title{Summarizing split_df information}
\usage{
\method{summary}{split_df}(object, ...)
}
\arguments{
\item{object}{an object of class "split_df", usually, a result of a call to split_df().}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary method for "split_df" class.
}
\details{
summary.split_df provides information on the number of two split data sets, minority class and majority class.
}
\examples{
library(dplyr)

# Credit Card Default Data
head(ISLR::Default)

# Generate data for the example
sb <- ISLR::Default \%>\%
  split_by(default)

sb
summary(sb)

}
